Bootloader first version
Created a bootloader for V2 micro:bits that should work with Finch 2, and likely Hummingbird Bit and Hatchling. This version seems to have some issues with serial communication - I need to send bytes slowly (1 every few ms), but it does seem to reliably program the firmware onto the BirdBrain products.